+++ flutter_readium/android/src/main/kotlin/dk/nota/flutterreadium/FlutterEpubPreferences.kt
@@ -47,6 +47,7 @@ data class FlutterEpubPreferences(
     val wordSpacing: Double? = null,
     val blackAndWhiteComicMode: Boolean? = false,
     val disableSynchronization: Boolean? = false,
+    val syncPolicy: String? = null,
     val firstElementTopMargin: Int? = null,
     val preventMOColumnBreaks: Boolean? = true,
 ) : Configurable.Preferences<FlutterEpubPreferences> {
@@ -78,6 +79,7 @@ data class FlutterEpubPreferences(
             wordSpacing = other.wordSpacing ?: wordSpacing,
             blackAndWhiteComicMode = other.blackAndWhiteComicMode ?: blackAndWhiteComicMode,
             disableSynchronization = other.disableSynchronization ?: disableSynchronization,
+            syncPolicy = other.syncPolicy ?: syncPolicy,
             firstElementTopMargin = other.firstElementTopMargin ?: firstElementTopMargin,
             preventMOColumnBreaks = other.preventMOColumnBreaks ?: preventMOColumnBreaks,
         )