+++ flutter_readium/test/flutter_readium_test.dart
@@ -1,4 +1,5 @@
 import 'dart:async';
+import 'dart:typed_data';
 
 import 'package:flutter_readium/flutter_readium.dart';
 import 'package:flutter_test/flutter_test.dart';
@@ -138,6 +139,9 @@ class MockFlutterReadiumPlatform with MockPlatformInterfaceMixin implements Flut
 
   @override
   Future<void> setLogLevel(LogLevel level) async {}
+
+  @override
+  Future<Uint8List> getResourceBytes(String href) async => Uint8List(0);
 
   @override
   Future<List<TextSearchResult>> searchInPublication(String searchKey) async => [];